Check active network state, and bounce interfaces if necessary (#28)
Sometimes the configuration files written to a host may not change, but the
active network state does not match the requested configuration. This can
happen following a previous configuration failure, where the configuration
files were successfully written but bouncing one or more interfaces failed. We
can also use this check after bouncing the interfaces to check that the
configuration has been successfully applied.

Currently IP routing tables, rules and routes are not checked, as these are not
available as ansible facts. In future, a routing fact gathering module could be
built to gather the required facts.