Add ``become: true`` to some tasks
I don't like to execute ansible roles with a general ``become: yes``.
So I added ``become: true`` to some of your tasks until your role worked for me well.

Maybe you want to take it?

As far as I know it does not hurt to add the ``become: true`` option on some points. especially when your playbook needs root permissions and is usually executet as root!