Fix use of 'match' test as filter (#64)
We currently see the following warning on Ansible 2.8:

DEPRECATION WARNING]: Using tests as filters is deprecated. Instead of
using result|match use result is match. This feature will be removed in
version 2.9. Deprecation warnings can be disabled by setting
deprecation_warnings=False in ansible.cfg.

2.9 is here, and now fails on this.

This change switches from '|' to 'is' for the 'match' test.

Fixes: #44