Add workaround for invalid interface config in CentOS cloud images
CentOS 8 cloud images ship with an ifcfg file for ens3. This seems to be
a relic from the image build process, and causes the network service to
fail.

This change adds a workaround that removes this file, if there is no
ens3 interface on the system, and no ens3 interface is specified in the
role variables. The interface config to be removed is defined via a
'interfaces_workaround_centos_remove' variable.