Fix persistent network configuration on RHEL/CentOS 8 (#69)
* Fix persistent network configuration on RHEL/CentOS 8

On RHEL/CentOS 8, this role now installs the network-scripts RPM. This
RPM provides the legacy ifup / ifdown implementations that can bypass
NetworkManager. It also provides a network init.d script. However, this
script is not enabled by default.

Since we use NM_CONTROLLED=no, this means that a role invocation will
work fine, until the next reboot or until the DHCP lease expires (when
using DHCP), whichever comes first.

This change fixes the issue by enabling network.service.

Fixes: #68

* Add workaround for invalid interface config in CentOS cloud images

CentOS 8 cloud images ship with an ifcfg file for ens3. This seems to be
a relic from the image build process, and causes the network service to
fail.

This change adds a workaround that removes this file, if there is no
ens3 interface on the system, and no ens3 interface is specified in the
role variables. The interface config to be removed is defined via a
'interfaces_workaround_centos_remove' variable.