Allow unmanaged addresses on an interface
Sometimes it may be useful to not immediately assign an IP address to an
interface, but to allow another process to assign one. An example use
case is a virtual IP address dynamically added or by a process such as
keepalived.

This change adds support for this via an 'allowed_addresses' field in
Ethernet, bond or bridge interface configuration.

Fixes: #70