Reload VLAN interface on parent device change
On CentOS/RHEL systems, if the parent device of a VLAN interface goes down, the
VLAN interface will also go down, and lose any static routes assigned to it. If
the parent device is brought back up again, the VLAN interface will become
active again, but any static routes previously assigned to the VLAN interface
are not reinstated.

This change fixes the issue by adding VLAN interfaces to the list of changed
interfaces, if their parent device is changed.

Fixes: #86