Restart bond interface after all members restart
On CentOS/RHEL systems, if all members in a bond go down, the bond interface
will also go down. If the members are brought back up again, the bond interface
does not automatically come back up. Therefore, if there is some change to the
configuration of bond members, the bond may be left in an inactive state.

This change fixes the issue by adding bond interfaces to the list of changed
interfaces, if all of their members have changed.

Fixes: #88