Restart bridge interface after all ports restart
On CentOS/RHEL 8 systems, if all ports in a bridge go down, the bridge
interface will also go down. If the ports are brought back up again, the bridge
interface does not automatically come back up. Therefore, if there is some
change to the configuration of all bridge ports, the bridge may be left in an
inactive state.

This change fixes the issue by adding bridge interfaces to the list of changed
interfaces, if all of their ports have changed.

Fixes: #90