RedHat: move network service start after interface config
The CentOS cloud images may include network interface files used when the image
was built, which may be invalid. This role already has a workaround to remove
them. It only removes files for interfaces that are not managed by this role.
This ensures that the role is idempotent. However, if an invalid interface file
exists for an interface that is managed by this role, it may prevent the
network service from starting, resulting in Ansible failing in the following
task:

  RedHat | ensure network service is started and enabled

This change moves the start of the network service after configuring
interfaces. This should overwrite any invalid configuration files of interfaces
managed by this role. This may result in some interfaces being bounced again in
handlers.

Fixes: #119