Move distro defaults into vars/distro
Putting these nested structures into vars allows us to load only the
distribution specific values that we need.

This uses a first_found lookup instead of a with_first_found loop so that the
'paths' parameter can be used. This ensures that only vars from the role are
included, and not vars from a parent calling role. This can happen when a
parent has a higher priority vars file available for inclusion than the role
it calls.