Policy-based routing for NetworkManager
Adding the option to send rules as a dictionary, we seem to need the
table_id when using nmconnections. As a dict, it's atleast feasible
to try looking up an id from a name in the templates.

We're not attempting to find the table name from a string in this
iteration. So if using nmconnections the table name needs to be
swapped for id.

It looks like the older network-scripts rules on EL-systems start
priority at 32765 and decrement as more are added, so we're mimicking
that behavior. It is also possible to override with a priority added
to the dict rule.

Also making old rule_RedHat.j2 template compatible with dicts.

Basing this on previous templates and documentation [1]
[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html/configuring_and_managing_networking/configuring-policy-based-routing-to-define-alternative-routes_configuring-and-managing-networking